Allow read-only connections if no password
If there is no password available let the connection to the wikis get setup, but do not attempt to login. This will allow all unit tests that do not write to still succeed, and then we can try to detect the state we are in and skip unit tests that may require writing (and limit them as much as possible).